community-rides-bot/src/main.py
"""
Community Rides — Telegram Bot (boilerplate).

This file is intentionally abstract. It defines:
    - environment/config loading
    - conversation state constants
    - handler function signatures with docstrings
    - application bootstrap shape

All business logic (storage, validation, formatting, posting to channel)
must be implemented by contributors inside the handlers.
"""

import os
import logging
from typing import Optional
from dotenv import load_dotenv

from telegram import Update
from telegram.ext import (
    Application,
    CommandHandler,
    ContextTypes,
    ConversationHandler,
    MessageHandler,
    filters,
)

# Storage API import.
import storage

load_dotenv()

TOKEN: Optional[str] = os.getenv("TELEGRAM_BOT_TOKEN")
CHANNEL_ID: Optional[str] = os.getenv("TELEGRAM_CHANNEL_ID")

logging.basicConfig(
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    level=logging.INFO,
)
logger = logging.getLogger("community-ride-bot")

# Conversation states (wizard).
FROM, TO, CAPACITY, TIME_RANGE, COMMENT, CONFIRM = range(6)

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """
    Send a short introduction and list available commands.

    For example:
        Explain purpose (drivers publish rides, bot posts to channel).
        List commands: /create, /update, /list, /list_all, /cancel.
    """
    pass


async def create_ride(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    """
    Start the ride creation wizard.

    Example flow:
        - Ask for "from" (start location).
        - Set up any per-user state in context.user_data if needed.
        - Return the next conversation state (FROM).
    """
    return FROM


async def from_received(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    """
    Capture start location and prompt for destination.

    Example flow:
        - Read and normalize free-text start location.
        - Store in context.user_data["from_location"].
        - Return TO state.
    """
    return TO


async def to_received(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    """
    Capture destination and prompt for capacity.

    Example flow:
        - Read and normalize free-text destination.
        - Store in context.user_data["to_location"].
        - Return CAPACITY state.
    """
    return CAPACITY


async def capacity_received(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    """
    Capture capacity (positive integer) and prompt for time range.

    Example flow:
        - Validate capacity (>= 1).
        - Store in context.user_data["capacity"].
        - Return TIME_RANGE state.
    """
    return TIME_RANGE


async def time_range_received(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    """
    Capture free-text time window and prompt for optional comment.

    Example flow:
        - Store in context.user_data["time_range"].
        - Return COMMENT state.
    """
    return COMMENT


async def comment_received(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    """
    Capture optional comment and present a confirmation summary.

    Example flow:
        - Store in context.user_data["comment"] (may be empty).
        - Build a summary of all fields.
        - Ask user to reply "yes" or "no" (confirmation).
        - Return CONFIRM state.
    """
    return CONFIRM


async def confirm_received(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    """
    Finalize ride creation if user confirms, otherwise cancel.

    Example flow on "yes":
        - Persist via storage.create_ride()
        - Post an announcement to CHANNEL_ID.
        - End the conversation.

    Example flow on "no":
        - Reply that creation is cancelled, end the conversation.

    Return:
        - ConversationHandler.END to stop the wizard, or another state to retry.
    """
    return ConversationHandler.END


async def list_active(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """
    List user's active rides (e.g. not expired/cancelled).

    Expected behavior:
        - Retrieve rides via storage.fetch_rides(user_id, active_only=True).
        - Format a readable list.
    """
    pass


async def list_all(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """
    List all rides created by the user (history).

    Expected behavior:
        - Retrieve rides via storage.fetch_rides(user_id, active_only=False).
        - Format and send as text (consider pagination later).
    """
    pass


async def delete_ride(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """
    Delete or deactivate a ride by its ID, owned by the current user.

    Example flow:
        - Parse /delete <ride_id>.
        - Call storage.delete_ride(user_id, ride_id, deactivate_only=True) by default.
        - Reply with success/failure message.
    """
    pass


async def cancel(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    """
    Abort the current wizard flow.

    Expected behavior:
        - Inform the user that the operation was cancelled.
        - Return ConversationHandler.END.
    """
    return ConversationHandler.END


def build_application(token: str) -> Application:
    """
    Create and configure the Application instance.

    Notes:
        - Register command and conversation handlers here.
        - Keep this function free of business logic.
    """
    app = Application.builder().token(token).build()

    conv = ConversationHandler(
        entry_points=[CommandHandler("create", create_ride)],
        states={
            FROM: [MessageHandler(filters.TEXT & ~filters.COMMAND, from_received)],
            TO: [MessageHandler(filters.TEXT & ~filters.COMMAND, to_received)],
            CAPACITY: [MessageHandler(filters.TEXT & ~filters.COMMAND, capacity_received)],
            TIME_RANGE: [MessageHandler(filters.TEXT & ~filters.COMMAND, time_range_received)],
            COMMENT: [MessageHandler(filters.TEXT & ~filters.COMMAND, comment_received)],
            CONFIRM: [MessageHandler(filters.TEXT & ~filters.COMMAND, confirm_received)],
        },
        fallbacks=[CommandHandler("cancel", cancel)],
        allow_reentry=True,
    )

    app.add_handler(CommandHandler("start", start))
    app.add_handler(CommandHandler("list", list_active))
    app.add_handler(CommandHandler("list_all", list_all))
    app.add_handler(CommandHandler("delete", delete_ride))
    app.add_handler(conv)

    return app


def main() -> None:
    """
    Entry point.
    """
    if not TOKEN:
        raise RuntimeError("Please set TELEGRAM_BOT_TOKEN")
    if not CHANNEL_ID:
        raise RuntimeError("Please set TELEGRAM_CHANNEL_ID")

    app = build_application(TOKEN)

    logger.info("Bot started.")
    app.run_polling()


if __name__ == "__main__":
    main()
