community-rides-bot/requirements.txt
python-telegram-bot>=21.5,<22
python-dotenv>=1.0,<2
